requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
torch==2.3.0+cu121
transformers==4.41.0
openai-whisper==20231117
llama-cpp-python==0.2.57
opencv-python==4.9.0.80
Pillow==10.3.0
scipy==1.13.0
aiosqlite==0.20.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
cryptography==42.0.5
python-multipart==0.0.9
pytest==8.2.0
